Introduce `is_with_tx_range` in DomainCoreApi
This commit is a pure refactoring to make the tx range check reusable in
domain-block-preprocessor later.

The sortition module in domain-operator is now not useful and deleted.
